Restructure files and path
